add exceptions

+++ src/main/java/com/nhnacademy/resident/service/impl/DefaultFamilyRelationshipService.java
@@ -4,6 +4,8 @@ import com.nhnacademy.resident.domain.request.RelationshipCreateRequest;
 import com.nhnacademy.resident.domain.request.RelationshipModifyRequest;
 import com.nhnacademy.resident.entity.FamilyRelationship;
 import com.nhnacademy.resident.entity.Resident;
+import com.nhnacademy.resident.exception.FamilyRelationshipNotFoundException;
+import com.nhnacademy.resident.exception.ResidentNotFoundException;
 import com.nhnacademy.resident.repository.FamilyRelationshipRepository;
 import com.nhnacademy.resident.repository.ResidentRepository;
 import com.nhnacademy.resident.service.FamilyRelationshipService;
@@ -27,6 +29,8 @@ public class DefaultFamilyRelationshipService implements FamilyRelationshipServi
         Optional<Resident> baseResident = residentRepository.findById(serialNumber);
         Optional<Resident> familyResident = residentRepository.findById(request.getFamilySerialNumber());
 
+        if (baseResident.isEmpty() || familyResident.isEmpty()) throw new ResidentNotFoundException();
+
         FamilyRelationship familyRelationship = new FamilyRelationship(new FamilyRelationship.Pk(serialNumber, request.getFamilySerialNumber()),
                 baseResident.get(), familyResident.get(), request.getRelationship());
         familyRelationshipRepository.save(familyRelationship);
@@ -36,10 +40,10 @@ public class DefaultFamilyRelationshipService implements FamilyRelationshipServi
     @Override
     public FamilyRelationship modifyFamilyRelationship(Long serialNumber, Long familySerialNumber, RelationshipModifyRequest request) {
         Optional<FamilyRelationship> familyRelationship = familyRelationshipRepository.findById(new FamilyRelationship.Pk(serialNumber, familySerialNumber));
+        if (familyRelationship.isEmpty()) throw new FamilyRelationshipNotFoundException();
+
         familyRelationship.get().setRelationshipCode(request.getRelationship());
-
         familyRelationshipRepository.updateRelationship(serialNumber, familySerialNumber, request.getRelationship());
-
         return familyRelationship.get();
     }
 
@@ -47,6 +51,8 @@ public class DefaultFamilyRelationshipService implements FamilyRelationshipServi
     public FamilyRelationship removeFamilyRelationship(Long serialNumber, Long familySerialNumber) {
         FamilyRelationship.Pk pk = new FamilyRelationship.Pk(serialNumber, familySerialNumber);
         Optional<FamilyRelationship> familyRelationship = familyRelationshipRepository.findById(pk);
+        if (familyRelationship.isEmpty()) throw new FamilyRelationshipNotFoundException();
+
         familyRelationshipRepository.deleteById(pk);
         return familyRelationship.get();
     }
