+++ src/main/java/com/nhnacademy/resident/service/impl/DefaultHouseholdService.java
@@ -4,6 +4,8 @@ import com.nhnacademy.resident.domain.request.HouseholdCreateRequest;
 import com.nhnacademy.resident.entity.Household;
 import com.nhnacademy.resident.entity.HouseholdComposition;
 import com.nhnacademy.resident.entity.Resident;
+import com.nhnacademy.resident.exception.HouseholdNotFoundException;
+import com.nhnacademy.resident.exception.ResidentNotFoundException;
 import com.nhnacademy.resident.repository.HouseholdCompositionRepository;
 import com.nhnacademy.resident.repository.HouseholdRepository;
 import com.nhnacademy.resident.repository.ResidentRepository;
@@ -28,6 +30,8 @@ public class DefaultHouseholdService implements HouseholdService {
     @Override
     public Household createHousehold(HouseholdCreateRequest request) {
         Optional<Resident> resident = residentRepository.findById(request.getResidentSerialNumber());
+        if (resident.isEmpty()) throw new ResidentNotFoundException();
+
         Household household = new Household(request.getSerialNumber(), resident.get(), request.getCompositionDate(), request.getCompositionReasonCode(), request.getCurrentAddress());
         householdRepository.save(household);
         householdCompositionRepository.save(new HouseholdComposition(new HouseholdComposition.Pk(request.getSerialNumber(), request.getResidentSerialNumber()),
@@ -38,6 +42,8 @@ public class DefaultHouseholdService implements HouseholdService {
     @Override
     public Household removeHousehold(Long householdSerialNumber) {
         Optional<Household> household = householdRepository.findById(householdSerialNumber);
+        if (household.isEmpty()) throw new HouseholdNotFoundException();
+
         householdRepository.deleteById(householdSerialNumber);
         householdCompositionRepository.deleteAllByHousehold(household.get());
         return household.get();
